separate metrics from operator, use only structs from go-metrics
