+++ api/go.mod
@@ -3,7 +3,7 @@ module github.com/hashicorp/nomad/api
 go 1.12
 
 require (
-	github.com/armon/go-metrics v0.3.4
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/docker/go-units v0.3.3
 	github.com/gorilla/websocket v1.4.1
 	github.com/hashicorp/cronexpr v1.1.0
@@ -12,5 +12,6 @@ require (
 	github.com/kr/pretty v0.1.0
 	github.com/mitchellh/go-testing-interface v1.0.0
 	github.com/stretchr/testify v1.5.1
+	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
 	gopkg.in/yaml.v2 v2.2.8 // indirect
 )
